requirements.txt
django
djangorestframework
markdown      
django-filter 
uvicorn
whitenoise
faker
django-cors-headers